static_gtfs/requirements.txt
# General dependencies
requests
python-dotenv
# Database dependencies
psycopg  # PostgreSQL driver for psycopg3
# Data processing
pandas
gtfs-realtime-bindings
